refactor: break out commands to separate files
